# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/doctors/{id}": {
      "get": {
        "operationId": "DoctorsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorDto"
                }
              }
            }
          },
          "404": {
            "description": "Doctor not found"
          },
          "503": {
            "description": "Error finding doctor"
          }
        },
        "tags": [
          "Doctors"
        ]
      }
    },
...
    "/appointments/{id}": {
      "get": {
        "operationId": "AppointmentsController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentDto"
                }
              }
            }
          },
          "404": {
            "description": "Appointment not found"
          },
          "503": {
            "description": "Error finding appointment"
          }
        },
        "tags": [
          "Appointments"
        ]
      }
    },
    "/appointments/doctors/{doctorId}": {
      "get": {
        "operationId": "AppointmentsController_findByDoctorId",
        "parameters": [
          {
            "name": "doctorId",
            "required": true,
            "in": "path",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppointmentDto"
                  }
                }
              }
            }
          },
          "503": {
            "description": "Error finding appointments"
          }
        },
        "tags": [
          "Appointments"
        ]
      }
    },
    "/appointments/patients/{patientId}": {
      "get": {
        "operationId": "AppointmentsController_findByPatientId",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppointmentDto"
                  }
                }
              }
            }
          },
          "503": {
            "description": "Error finding appointments"
          }
        },
        "tags": [
          "Appointments"
        ]
      }
    },
    "/appointments": {
      "post": {
        "operationId": "AppointmentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentDto"
                }
              }
            }
          },
          "403": {
            "description": "Timeslot is already taken or Doctor does not match"
          },
          "404": {
            "description": "Doctor, Patient or Timeslot not found"
          },
          "503": {
            "description": "Error creating appointment"
          }
        },
        "tags": [
          "Appointments"
        ]
      }
    },
    "/appointments/result": {
      "post": {
        "operationId": "AppointmentsController_createResult",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddAppointmentResultDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResultDto"
                }
              }
            }
          },
          "404": {
            "description": "Appointment not found"
          },
          "503": {
            "description": "Error creating appointment result"
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "Appointments"
        ]
      },
      "patch": {
        "operationId": "AppointmentsController_updateResult",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAppointmentResultDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResultDto"
                }
              }
            }
          },
          "403": {
            "description": "User is not in appointment "
          },
          "404": {
            "description": "Appointment not found"
          },
          "503": {
            "description": "Error updating appointment result"
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "Appointments"
        ]
      }
    },
    "/appointments/decline": {
      "post": {
        "operationId": "AppointmentsController_declineAppointment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeclineAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentDto"
                }
              }
            }
          },
          "404": {
            "description": "Appointment not found"
          },
          "503": {
            "description": "Error declining appointment"
          }
        },
        "tags": [
          "Appointments"
        ]
      }
    },
...
    "/patients/{id}": {
      "get": {
        "operationId": "PatientsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientDto"
                }
              }
            }
          },
          "404": {
            "description": "Patient not found"
          },
          "503": {
            "description": "Error finding patient"
          }
        },
        "tags": [
          "Patients"
        ]
      }
    },
...
  },
  "info": {
    "title": "Appointments service API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "DoctorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "phone_number": {
            "type": "string"
          },
          "licence_number": {
            "type": "string"
          },
          "specialization_id": {
            "type": "number"
          },
          "user_id": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "phone_number",
          "licence_number",
          "specialization_id",
          "user_id"
        ]
      },
...
      "PatientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "user_id": {
            "type": "number"
          },
          "date_of_birth": {
            "format": "date-time",
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_id",
          "date_of_birth",
          "gender",
          "address",
          "created_at",
          "updated_at"
        ]
      },
...
      "TimeslotDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "start_time": {
            "format": "date-time",
            "type": "string"
          },
          "end_time": {
            "format": "date-time",
            "type": "string"
          },
          "is_available": {
            "type": "boolean"
          },
          "doctor_id": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "start_time",
          "end_time",
          "is_available",
          "doctor_id"
        ]
      },
      "AppointmentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "full_name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone_number": {
            "type": "string"
          },
          "patient_insurance_number": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "timeslot_id": {
            "type": "number"
          },
          "doctor_id": {
            "type": "number"
          },
          "patient": {
            "$ref": "#/components/schemas/PatientDto"
          },
          "doctor": {
            "$ref": "#/components/schemas/DoctorDto"
          },
          "timeslot": {
            "$ref": "#/components/schemas/TimeslotDto"
          }
        },
        "required": [
          "id",
          "full_name",
          "email",
          "phone_number",
          "patient_insurance_number",
          "reason",
          "timeslot_id",
          "doctor_id",
          "patient",
          "doctor",
          "timeslot"
        ]
      },
      "CreateAppointmentDto": {
        "type": "object",
        "properties": {
          "full_name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone_number": {
            "type": "string"
          },
          "patient_insurance_number": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "timeslot_id": {
            "type": "number"
          },
          "doctor_id": {
            "type": "number"
          }
        },
        "required": [
          "full_name",
          "email",
          "phone_number",
          "patient_insurance_number",
          "reason",
          "timeslot_id",
          "doctor_id"
        ]
      },
      "AddAppointmentResultDto": {
        "type": "object",
        "properties": {
          "appointmentId": {
            "type": "number"
          },
          "diagnosis": {
            "type": "string"
          },
          "recommendations": {
            "type": "string"
          }
        },
        "required": [
          "appointmentId",
          "diagnosis",
          "recommendations"
        ]
      },
      "AppointmentResultDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "diagnosis": {
            "type": "string"
          },
          "recommendations": {
            "type": "string"
          },
          "appointment_id": {
            "type": "number"
          },
          "appointment": {
            "$ref": "#/components/schemas/AppointmentDto"
          }
        },
        "required": [
          "id",
          "diagnosis",
          "recommendations",
          "appointment_id",
          "appointment"
        ]
      },
      "UpdateAppointmentResultDto": {
        "type": "object",
        "properties": {
          "appointmentId": {
            "type": "number"
          },
          "diagnosis": {
            "type": "string"
          },
          "recommendations": {
            "type": "string"
          }
        },
        "required": [
          "appointmentId"
        ]
      },
      "DeclineAppointmentDto": {
        "type": "object",
        "properties": {
          "appointmentId": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "appointmentId"
        ]
      },
...
    }
  }
}